Minimize AnalogClock class code size
A single AnalogClock class code size was getting too large and
was resulting in odd execution issues with working code.
Stripped AnalogClock to just the time and clock drawing code.
Created example codes for supported board. Each board inherets
the AnalogClock class with the following requirements:
- __init__ will define colors, create portal, create display, and
call network
- networkConnect will contain any code needed to create and use
a network interface and get network time. The MacroPad uses this
code to querry the attached RTC.
- AnalogClock.display has an optional wait parameter to force a sleep
between screen updates. The Magtag_AnalogClock uses this as an example
both in the code.py and the implemented display() call.